+++ src/flavia/tools/content/__init__.py
@@ -5,6 +5,11 @@ from .query_catalog import QueryCatalogTool
 from .get_summary import GetSummaryTool
 from .refresh_catalog import RefreshCatalogTool
 from .search_chunks import SearchChunksTool
+from .add_online_source import AddOnlineSourceTool
+from .fetch_online_source import FetchOnlineSourceTool
+from .convert_pdf import ConvertPdfTool
+from .convert_office import ConvertOfficeTool
+from .transcribe_media import TranscribeMediaTool
 
 from ..registry import register_tool
 
@@ -14,6 +19,11 @@ register_tool(QueryCatalogTool())
 register_tool(GetSummaryTool())
 register_tool(RefreshCatalogTool())
 register_tool(SearchChunksTool())
+register_tool(AddOnlineSourceTool())
+register_tool(FetchOnlineSourceTool())
+register_tool(ConvertPdfTool())
+register_tool(ConvertOfficeTool())
+register_tool(TranscribeMediaTool())
 
 __all__ = [
     "AnalyzeImageTool",
@@ -21,4 +31,9 @@ __all__ = [
     "GetSummaryTool",
     "RefreshCatalogTool",
     "SearchChunksTool",
+    "AddOnlineSourceTool",
+    "FetchOnlineSourceTool",
+    "ConvertPdfTool",
+    "ConvertOfficeTool",
+    "TranscribeMediaTool",
 ]